# feat: optimize all CGW setup and keygen using wNAF precomputations