The setup and keygen of Chen-Gay-Wee schemes perform a significant amount of exponentiations based on their respective generators. Precomputing tables helps speed this up and is supported generically by the `group` crate. This does require the `alloc` crate though, but so do multi-miller loops, which we are already using in the decryption.

On my laptop:
```
kem_cgw_fo setup        time:   [5.8347 ms 5.9195 ms 6.0213 ms]
                        change: [-29.972% -27.270% -24.835%] (p = 0.00 < 0.05)
                        Performance has improved.

kem_cgw_fo extract      time:   [3.0686 ms 3.1267 ms 3.1946 ms]
                        change: [-54.603% -53.314% -52.080%] (p = 0.00 < 0.05)
                        Performance has improved.

kem_cgw_kv setup        time:   [7.1707 ms 7.7353 ms 8.3904 ms]
                        change: [-28.164% -21.793% -14.761%] (p = 0.00 < 0.05)
                        Performance has improved.

kem_cgw_kv extract      time:   [4.5529 ms 4.6296 ms 4.7187 ms]
                        change: [-62.013% -59.866% -57.652%] (p = 0.00 < 0.05)
                        Performance has improved.
                        
ibe_cgw setup           time:   [5.8650 ms 5.9957 ms 6.1463 ms]
                        change: [-36.213% -32.878% -29.331%] (p = 0.00 < 0.05)
                        Performance has improved.

ibe_cgw extract         time:   [3.0603 ms 3.1170 ms 3.1846 ms]
                        change: [-55.153% -53.607% -52.057%] (p = 0.00 < 0.05)
                        Performance has improved.
```