@leonbotros 2022-11-17
Theoretically the base table does not have to use dynamic memory usage, and instead a constant. But `group` does not allow this pattern yet.